perf(scheduler,exporter): use KeyedController for parallel per-object… (#649)
* perf(scheduler,exporter): use KeyedController for parallel per-object processing

Speed up scheduling and DB export by keying work queues on workspace and
resource identity so duplicate events coalesce while distinct objects run in
parallel. Consolidate exporter unit tests into one file per source module.


* fix(scheduler,exporter): repair keyed tests and align exporter workers to 8

Fix scheduler test compile error by using client.Client, complete parallel
workspace coalesce assertions, and match exporter per-GVK concurrency with
scheduler and syncer.


---------
