+++ SaFE/resource-manager/pkg/exporter/resource_exporter_test.go
@@ -7,7 +7,9 @@ package exporter
 
 import (
 	"context"
+	"sync"
 	"testing"
+	"time"
 
 	"gotest.tools/assert"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -40,7 +42,7 @@ func newTestExporter(t *testing.T, objs ...*v1.Workload) *ResourceExporter {
 		Client: builder.Build(),
 		gvk:    v1.SchemeGroupVersion.WithKind(v1.WorkloadKind),
 	}
-	exp.Controller = commonctrl.NewController[types.NamespacedName](exp, 1)
+	exp.KeyedController = commonctrl.NewKeyedController[types.NamespacedName](exp, namespacedNameKey, nil, exporterWorkers)
 	return exp
 }
 
@@ -61,7 +63,6 @@ func TestExporterGetObjectNotFound(t *testing.T) {
 
 func TestExporterDoNotFound(t *testing.T) {
 	exp := newTestExporter(t)
-	// Missing object -> ignored, no error.
 	_, err := exp.Do(context.Background(), types.NamespacedName{Name: "missing"})
 	assert.NilError(t, err)
 }
@@ -78,7 +79,6 @@ func TestExporterDoAddsFinalizerAndHandles(t *testing.T) {
 	assert.NilError(t, err)
 	assert.Assert(t, handled)
 
-	// Finalizer should now be present.
 	obj, err := exp.getObject(context.Background(), types.NamespacedName{Name: "w1"})
 	assert.NilError(t, err)
 	found := false
@@ -94,7 +94,6 @@ func TestExporterStart(t *testing.T) {
 	exp := newTestExporter(t)
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
-	// start spawns worker goroutines; should not block or panic.
 	assert.NilError(t, exp.start(ctx))
 }
 
@@ -123,10 +122,56 @@ func TestExporterFinalizerHelpers(t *testing.T) {
 	assert.NilError(t, err)
 
 	assert.NilError(t, exp.addFinalizer(context.Background(), obj))
-	// Adding again is a no-op (finalizer already present).
 	assert.NilError(t, exp.addFinalizer(context.Background(), obj))
 
 	assert.NilError(t, exp.removeFinalizer(context.Background(), obj))
-	// Removing again is a no-op.
 	assert.NilError(t, exp.removeFinalizer(context.Background(), obj))
+}
+
+type exportCountHandler struct {
+	mu     sync.Mutex
+	counts map[string]int
+}
+
+func (h *exportCountHandler) Do(_ context.Context, nn types.NamespacedName) (ctrlruntime.Result, error) {
+	h.mu.Lock()
+	h.counts[nn.String()]++
+	h.mu.Unlock()
+	return ctrlruntime.Result{}, nil
+}
+
+func (h *exportCountHandler) get(key string) int {
+	h.mu.Lock()
+	defer h.mu.Unlock()
+	return h.counts[key]
+}
+
+// TestExporterKeyedCoalesce verifies duplicate object events collapse to one export.
+func TestExporterKeyedCoalesce(t *testing.T) {
+	h := &exportCountHandler{counts: make(map[string]int)}
+	c := commonctrl.NewKeyedController[types.NamespacedName](h, namespacedNameKey, nil, 1)
+
+	nn := types.NamespacedName{Name: "w1", Namespace: "primus-safe"}
+	c.Add(nn)
+	c.Add(nn)
+	c.Add(nn)
+	assert.Equal(t, 1, c.GetQueueSize())
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	c.Run(ctx)
+	deadline := time.Now().Add(2 * time.Second)
+	for c.GetQueueSize() > 0 && time.Now().Before(deadline) {
+		time.Sleep(10 * time.Millisecond)
+	}
+	time.Sleep(50 * time.Millisecond)
+	cancel()
+	time.Sleep(50 * time.Millisecond)
+
+	assert.Equal(t, 1, h.get(nn.String()))
+}
+
+func TestNamespacedNameKey(t *testing.T) {
+	nn := types.NamespacedName{Name: "w1", Namespace: "primus-safe"}
+	assert.Equal(t, nn.String(), namespacedNameKey(nn))
 }
